separated transfer learning methods from preprocessing operators scripts
